sheriff valid
